+++ example/lib/main.dart
@@ -7,6 +7,8 @@ import 'package:instabug_flutter/Instabug.dart';
 import 'package:instabug_flutter/BugReporting.dart';
 import 'package:instabug_flutter/InstabugLog.dart';
 import 'package:instabug_flutter/Surveys.dart';
+import 'package:instabug_flutter/FeatureRequests.dart';
+import 'package:instabug_flutter/Chats.dart';
 
 void main() => runApp(MyApp());
 
@@ -123,8 +125,13 @@ class _MyAppState extends State<MyApp> {
     // Surveys.getAvailableSurveys(getSurveys);
     // Surveys.showSurveyIfAvailable();
     // Surveys.setShouldShowWelcomeScreen(true);
-    Surveys.showSurvey("BHJI1iaKYhr4CYHHcUAaTg");
+    //Surveys.showSurvey("BHJI1iaKYhr4CYHHcUAaTg");
     //BugReporting.showWithOptions(ReportType.bug, <InvocationOption>[InvocationOption.emailFieldHidden]);
+    // FeatureRequests.setEmailFieldRequired(false, [ActionType.allActions]);
+    // FeatureRequests.show();
+    Chats.setEnabled(true);
+    Chats.show();
+    
   }
 
   void invokeWithMode() {