+++ test/instabug_flutter_test.dart
@@ -8,6 +8,8 @@ import 'package:instabug_flutter/InstabugLog.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:instabug_flutter/Surveys.dart';
+import 'package:instabug_flutter/FeatureRequests.dart';
+import 'package:instabug_flutter/Chats.dart';
 
 void main() {
 
@@ -528,6 +530,44 @@ test('startWithToken:invocationEvents: Test', () async {
     expect(log, <Matcher>[
       isMethodCall('hasRespondedToSurveyWithToken:',
       arguments: args,
+      )
+    ]);
+  });
+
+  test('showFeatureRequests Test', () async {
+    FeatureRequests.show();
+    expect(log, <Matcher>[
+      isMethodCall('showFeatureRequests'
+     )
+    ]);
+  });
+
+  test('setEmailFieldRequiredForFeatureRequests:forAction: Test', () async {
+    bool isEmailFieldRequired = false;
+    final List<dynamic> args = <dynamic>[isEmailFieldRequired, <String>[ActionType.allActions.toString()]];
+    FeatureRequests.setEmailFieldRequired(isEmailFieldRequired, [ActionType.allActions]);
+    expect(log, <Matcher>[
+      isMethodCall('setEmailFieldRequiredForFeatureRequests:forAction:',
+      arguments: args,
+     )
+    ]);
+  });
+
+  test('showChats Test', () async {
+    Chats.show();
+    expect(log, <Matcher>[
+      isMethodCall('showChats'
+     )
+    ]);
+  });
+
+  test('setChatsEnabled: Test', () async {
+    bool isEnabled = false;
+    final List<dynamic> args = <dynamic>[isEnabled];
+    Chats.setEnabled(isEnabled);
+    expect(log, <Matcher>[
+      isMethodCall('setChatsEnabled:',
+        arguments: args,
       )
     ]);
   });